refactor(api): use constants for readability
