+++ src/github/api/issues/github-api-issues.service.ts
@@ -1,3 +1,4 @@
+import { GITHUB_API_ISSUES_QUERY } from '@github/api/issues/github-api-issues-query';
 import { GITHUB_ISSUES_PER_PAGE } from '@github/api/issues/github-issues-per-page';
 import { IGithubApiIssues } from '@github/api/issues/interfaces/github-api-issues.interface';
 import { OctokitService } from '@github/octokit/octokit.service';
@@ -13,34 +14,12 @@ export class GithubApiIssuesService {
     LoggerService.info(`Fetching the issues from GitHub...`);
 
     return OctokitService.getOctokit()
-      .graphql<IGithubApiIssues>(
-        `
-        query MyQuery($owner: String!, $repository: String!, $issuesPerPage: Int!, $afterCursor: String) {
-          repository(name: $repository, owner: $owner) {
-            issues(orderBy: {field: UPDATED_AT, direction: DESC}, states: OPEN, first: $issuesPerPage, after: $afterCursor) {
-              pageInfo {
-                hasNextPage
-              }
-              totalCount
-              nodes {
-                locked
-                createdAt
-                number
-                updatedAt
-                url
-                id
-              }
-            }
-          }
-        }
-      `,
-        {
-          afterCursor: fromPageId,
-          issuesPerPage: GithubApiIssuesService.issuesPerPage,
-          owner: context.repo.owner,
-          repository: context.repo.repo,
-        }
-      )
+      .graphql<IGithubApiIssues>(GITHUB_API_ISSUES_QUERY, {
+        afterCursor: fromPageId,
+        issuesPerPage: GithubApiIssuesService.issuesPerPage,
+        owner: context.repo.owner,
+        repository: context.repo.repo,
+      })
       .then((response: Readonly<IGithubApiIssues>): IGithubApiIssues => {
         // Only log the first time (when we do not have some pagination yet)
         if (!fromPageId) {
